Moves validator from lib to lib/pwned
